+++ src/SpiNRG.h
@@ -34,9 +34,9 @@ public:
          * @param size
          * @param rxData Leave nullptr if not interested.
          */
-        void transmit8 (uint8_t const *txData, uint16_t size, uint8_t *rxData = nullptr, size_t bogoDelay = 0);
+        void transmit8 (uint8_t const *txData, size_t size, uint8_t *rxData = nullptr, size_t bogoDelay = 0);
 
-        void receive8 (uint8_t *rxData, uint16_t size, size_t bogoDelay = 0);
+        void receive8 (uint8_t *rxData, size_t size, size_t bogoDelay = 0);
 
         /**
          * @brief Transmit one byte, and return what has been sent back.
